Refacto tests
