+++ tests/Functional/CustomerDirectDebitValidationTest.php
@@ -20,7 +20,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace tests;
+namespace Digitick\Sepa\Tests;
 
 use Digitick\Sepa\DomBuilder\CustomerDirectDebitTransferDomBuilder;
 use Digitick\Sepa\Exception\InvalidTransferFileConfiguration;
@@ -55,8 +55,8 @@ class CustomerDirectDebitValidationTest extends TestCase
      */
     public function testSanity($schema)
     {
-        $this->dom->load(__DIR__ . '/' . $schema . '.xml');
-        $validated = $this->dom->schemaValidate(__DIR__ . '/' . $schema . '.xsd');
+        $this->dom->load(__DIR__ . '/../fixtures/' . $schema . '.xml');
+        $validated = $this->dom->schemaValidate(__DIR__ . '/../fixtures/' . $schema . '.xsd');
         $this->assertTrue($validated);
     }
 
@@ -92,7 +92,7 @@ class CustomerDirectDebitValidationTest extends TestCase
         $xml = $domBuilder->asXml();
         $this->dom->loadXML($xml);
 
-        $validated = $this->dom->schemaValidate(__DIR__ . '/' . $schema . '.xsd');
+        $validated = $this->dom->schemaValidate(__DIR__ . '/../fixtures/' . $schema . '.xsd');
         $this->assertTrue($validated);
     }
 
@@ -128,7 +128,7 @@ class CustomerDirectDebitValidationTest extends TestCase
         $xml = $domBuilder->asXml();
         $this->dom->loadXML($xml);
 
-        $validated = $this->dom->schemaValidate(__DIR__ . '/' . $schema . '.xsd');
+        $validated = $this->dom->schemaValidate(__DIR__ . '/../fixtures/' . $schema . '.xsd');
         $this->assertTrue($validated);
 
         $xpathDoc = new \DOMXPath($this->dom);
@@ -291,7 +291,7 @@ class CustomerDirectDebitValidationTest extends TestCase
         $xml = $domBuilder->asXml();
         $this->dom->loadXML($xml);
 
-        $validated = $this->dom->schemaValidate(__DIR__ . '/' . $scenario['pain'] . '.xsd');
+        $validated = $this->dom->schemaValidate(__DIR__ . '/../fixtures/' . $scenario['pain'] . '.xsd');
         $this->assertTrue($validated);
     }
