+++ tests/Unit/DomBuilder/DomBuilderFactoryTest.php
@@ -20,7 +20,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace tests\Unit;
+namespace Digitick\Sepa\Tests\Unit\DomBuilder;
 
 use Digitick\Sepa\DomBuilder\DomBuilderFactory;
 use Digitick\Sepa\GroupHeader;
